+++ src/Command/ExportCommand.php
@@ -32,7 +32,7 @@ class ExportCommand extends Command
      * @param \Cake\Console\ConsoleOptionParser $parser The parser to be defined
      * @return \Cake\Console\ConsoleOptionParser
      */
-    protected function buildOptionParser(ConsoleOptionParser $parser)
+    protected function buildOptionParser(ConsoleOptionParser $parser): ConsoleOptionParser
     {
         return $parser->setDescription(__d('database_backup', 'Exports a database backup'))
             ->addOptions([
@@ -66,7 +66,7 @@ class ExportCommand extends Command
      * This command uses `RotateCommand` and `SendCommand`.
      * @param \Cake\Console\Arguments $args The command arguments
      * @param \Cake\Console\ConsoleIo $io The console io
-     * @return null|int The exit code or null for success
+     * @return int|null The exit code or null for success
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupShell#export
      * @uses DatabaseBackup\Command\RotateCommand::execute()
      * @uses DatabaseBackup\Command\SendCommand::execute()
@@ -74,7 +74,7 @@ class ExportCommand extends Command
      * @uses DatabaseBackup\Utility\BackupExport::export()
      * @uses DatabaseBackup\Utility\BackupExport::filename()
      */
-    public function execute(Arguments $args, ConsoleIo $io)
+    public function execute(Arguments $args, ConsoleIo $io): ?int
     {
         parent::execute($args, $io);
 
@@ -114,7 +114,8 @@ class ExportCommand extends Command
                     $RotateCommand->getOptionParser()->argumentNames()
                 ), $io);
             }
-        } catch (Exception $e) {
+        }
+        catch (Exception $e) {
             $io->error($e->getMessage());
             $this->abort();
         }