+++ src/Console/Command.php
@@ -29,9 +29,9 @@ class Command extends BaseCommand
      * Implement this method with your command's logic
      * @param \Cake\Console\Arguments $args The command arguments
      * @param \Cake\Console\ConsoleIo $io The console io
-     * @return null|int The exit code or null for success
+     * @return int|null The exit code or null for success
      */
-    public function execute(Arguments $args, ConsoleIo $io)
+    public function execute(Arguments $args, ConsoleIo $io): ?int
     {
         $config = $this->getConnection()->config();
         $driver = $this->getDriver($this->getConnection());