+++ src/TestSuite/DriverTestCase.php
@@ -69,7 +69,7 @@ abstract class DriverTestCase extends TestCase
      * Called before every test method
      * @return void
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();