+++ tests/TestCase/Utility/BackupImportTest.php
@@ -39,7 +39,7 @@ class BackupImportTest extends TestCase
      * Called before every test method
      * @return void
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();