+++ tests/TestCase/Utility/BackupManagerTest.php
@@ -14,9 +14,8 @@ namespace DatabaseBackup\Test\TestCase\Utility;
 
 use Cake\Core\Configure;
 use Cake\I18n\FrozenTime;
-use Cake\Mailer\Email;
 use Cake\ORM\Entity;
-use Cake\TestSuite\EmailAssertTrait;
+use Cake\TestSuite\EmailTrait;
 use DatabaseBackup\TestSuite\TestCase;
 use DatabaseBackup\Utility\BackupExport;
 use DatabaseBackup\Utility\BackupManager;
@@ -29,7 +28,7 @@ use Tools\Exception\NotWritableException;
  */
 class BackupManagerTest extends TestCase
 {
-    use EmailAssertTrait;
+    use EmailTrait;
 
     /**
      * @var \DatabaseBackup\Utility\BackupExport
@@ -45,7 +44,7 @@ class BackupManagerTest extends TestCase
      * Called before every test method
      * @return void
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
@@ -169,20 +168,18 @@ class BackupManagerTest extends TestCase
         $mimetype = mime_content_type($file);
         $to = 'recipient@example.com';
 
-        $instance = new BackupManager();
-        $this->_email = $this->invokeMethod($instance, 'getEmailInstance', [$file, $to]);
-        $this->assertInstanceof(Email::class, $this->_email);
+        $this->_email = $this->invokeMethod($this->BackupManager, 'getEmailInstance', [$file, $to]);
+        $this->BackupManager->send($file, $to);
 
-        $this->assertEmailFrom(Configure::read('DatabaseBackup.mailSender'));
-        $this->assertEmailTo($to);
-        $this->assertEmailSubject('Database backup ' . basename($file) . ' from localhost');
-        $this->assertEmailAttachmentsContains(basename($file), compact('file', 'mimetype'));
-        $this->assertArrayKeysEqual(['headers', 'message'], $this->BackupManager->send($file, $to));
+        $this->assertMailSentFrom(Configure::read('DatabaseBackup.mailSender'));
+        $this->assertMailSentTo($to);
+        $this->assertMailSentWith('Database backup ' . basename($file) . ' from localhost', 'subject');
+        $this->assertSame([basename($file) => compact('file', 'mimetype')], $this->_email->getAttachments());
 
         //With an invalid sender
-        @unlink($file);
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Invalid email set for "from". You passed "invalidSender".');
+        @unlink($file);
         Configure::write('DatabaseBackup.mailSender', 'invalidSender');
         $this->BackupManager->send($this->createBackup(), 'recipient@example.com');
     }