+++ tests/bootstrap.php
@@ -17,6 +17,7 @@ use Cake\Datasource\ConnectionManager;
 use Cake\Log\Log;
 use Cake\Mailer\Email;
 use Cake\Mailer\TransportFactory;
+use Cake\TestSuite\TestEmailTransport;
 
 require dirname(__DIR__) . '/vendor/autoload.php';
 
@@ -115,13 +116,8 @@ Log::setConfig('debug', [
 ]);
 
 $transportName = 'debug';
-$transportConfig = ['className' => 'Debug'];
-if (class_exists(TransportFactory::class)) {
-    TransportFactory::setConfig($transportName, $transportConfig);
-} else {
-    Email::setConfigTransport($transportName, $transportConfig);
-}
-Email::setConfig('default', ['transport' => $transportName, 'log' => true]);
+TransportFactory::setConfig($transportName, ['className' => TestEmailTransport::class]);
+Email::setConfig('default', ['transport' => $transportName]);
 
 Configure::write('pluginsToLoad', ['DatabaseBackup']);